requirements.txt
selenium
webdriver-manager
requests
Pillow
reportlab
